% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apidevice.R
\docType{class}
\name{ScDevice-class}
\alias{ScDevice-class}
\alias{ScDevice}
\title{Title}
\description{
Class to interface the SC API
}
\section{Fields}{

\describe{
\item{\code{id}}{numeric.}

\item{\code{json}}{ANY.}

\item{\code{data}}{ANY.}

\item{\code{names}}{ANY.}
}}

\section{Methods}{

\describe{
\item{\code{get_device_data(rollup = "1m", min_date = "", max_date = "")}}{Gets tabular data of the device
 Parameters
   rollup:
     String. Default: 1m
     See reference in https://developer.smartcitizen.me/#rollup-measurements
   min_date:
     String. Default: ''
     Minimum date to request to the API (from). Format YYYY-MM-DD
   max_date
     String. Default: ''
     Maximum date to request to the API (from). Format YYYY-MM-DD}

\item{\code{get_json()}}{Gets a json with information about the device}

\item{\code{get_names()}}{Gets a table with the names of the sensors in the device}
}}

